+++ skin/frontend/base/default/stripe_payments/js/stripe_payments.js
@@ -46,7 +46,7 @@ var initStripe = function(apiKey, securityMethod, callback)
 };
 
 var stripe = {
-    version: '1.1.8',
+    version: '1.1.9',
 
     // Properties
     billingInfo: null,
@@ -1162,33 +1162,6 @@ var initAdmin = function()
     {
         topBtns[i].onclick = stripe.placeAdminOrder;
     }
-};
-
-var beginApplePay = function()
-{
-    var paymentRequest = paramsApplePay;
-
-    var countryCode = stripe.getCountryCode();
-    if (countryCode && countryCode != paymentRequest.countryCode)
-    {
-        // In some cases with OSC modules, the country may change without having the payment form reloaded
-        paymentRequest.countryCode = countryCode;
-    }
-
-    var ession = Stripe.applePay.buildSession(paymentRequest, function(result, completion)
-    {
-        setStripeToken(result.token.id);
-
-        completion(ApplePaySession.STATUS_SUCCESS);
-
-        setApplePayToken(result.token);
-    },
-    function(error)
-    {
-        alert(error.message);
-    });
-
-    session.begin();
 };
 
 var setApplePayToken = function(token)